hw4/requirements.txt
addict==2.4.0
annotated-types==0.7.0
anyio==4.11.0
attrs==25.3.0
av==15.1.0
babel==2.17.0
blis==1.3.0
catalogue==2.0.10
certifi==2025.8.3
cffi==2.0.0
charset-normalizer==3.4.3
click==8.3.0
cloudpathlib==0.22.0
coloredlogs==15.0.1
confection==0.1.5
csvw==3.6.0
ctranslate2==4.6.0
curated-tokenizers==0.0.9
curated-transformers==0.1.1
cymem==2.0.11
dlinfo==2.0.0
docopt==0.6.2
en_core_web_sm @ https://github.com/explosion/spacy-models/releases/download/en_core_web_sm-3.8.0/en_core_web_sm-3.8.0-py3-none-any.whl#sha256=1932429db727d4bff3deed6b34cfc05df17794f4a52eeb26cf8928f7c1a0fb85
espeakng-loader==0.2.4
faster-whisper==1.2.0
filelock==3.19.1
flatbuffers==25.9.23
fsspec==2025.9.0
h11==0.16.0
hf-xet==1.1.10
httpcore==1.0.9
httpx==0.28.1
httpx-sse==0.4.1
huggingface-hub==0.35.3
humanfriendly==10.0
idna==3.10
isodate==0.7.2
Jinja2==3.1.6
joblib==1.5.2
jsonschema==4.25.1
jsonschema-specifications==2025.9.1
kokoro==0.9.4
langcodes==3.5.0
language-tags==1.2.0
language_data==1.3.0
loguru==0.7.3
marisa-trie==1.3.1
markdown-it-py==4.0.0
MarkupSafe==3.0.3
mcp==1.16.0
mdurl==0.1.2
misaki==0.9.4
mpmath==1.3.0
murmurhash==1.0.13
networkx==3.5
num2words==0.5.14
numpy==2.3.3
onnxruntime==1.23.0
packaging==25.0
phonemizer-fork==3.3.2
preshed==3.0.10
protobuf==6.32.1
pycparser==2.23
pydantic==2.11.9
pydantic-settings==2.11.0
pydantic_core==2.33.2
Pygments==2.19.2
pyparsing==3.2.5
python-dateutil==2.9.0.post0
python-dotenv==1.1.1
python-multipart==0.0.20
PyYAML==6.0.3
rdflib==7.2.1
referencing==0.36.2
regex==2025.9.18
requests==2.32.5
rfc3986==1.5.0
rich==14.1.0
rpds-py==0.27.1
safetensors==0.6.2
segments==2.3.0
shellingham==1.5.4
six==1.17.0
smart_open==7.3.1
sniffio==1.3.1
soundfile==0.13.1
spacy==3.8.7
spacy-curated-transformers==0.3.1
spacy-legacy==3.0.12
spacy-loggers==1.0.5
srsly==2.5.1
sse-starlette==3.0.2
starlette==0.48.0
sympy==1.14.0
termcolor==3.1.0
thinc==8.3.6
tokenizers==0.22.1
torch==2.8.0
tqdm==4.67.1
transformers==4.56.2
typer==0.19.2
typing-inspection==0.4.2
typing_extensions==4.15.0
uritemplate==4.2.0
urllib3==2.5.0
uvicorn==0.37.0
wasabi==1.1.3
weasel==0.4.1
wrapt==1.17.3
